requirements.txt
fastapi
uvicorn
yt-dlp
pydantic
httpx
python-telegram-bot==20.7
python-dotenv